+++ sdk/merchantPurchase/index.js
@@ -4,3 +4,5 @@ exports.MerchantPurchase = merchantPurchase.MerchantPurchase;
 exports.query = merchantPurchase.query;
 exports.get = merchantPurchase.get;
 exports.page = merchantPurchase.page;
+exports.create = merchantPurchase.create;
+exports.update = merchantPurchase.update;